feat: ping-1480 move process to helper
