+++ src/processes/registerv2-process.js
@@ -20,6 +20,7 @@ const LocationFunction = require('../databases/functions/location');
 const {addUserToTopicChannel} = require('./helper');
 const ProcessHelper = require('./helper');
 const {automateWelcomeMsgQueue} = require('../config');
+const {preAutomateWelcomeMsgProcess} = require('./helper/preAutoWelcomeMsg');
 
 const {syncFeedPerUserProcess} = require('../services/syncFeedPerUser');
 
@@ -62,30 +63,7 @@ const registerProcess = async (job, done) => {
 
     await ProcessHelper.followDefaultLocation(userId);
     const returnPrepopulatedDm = await ProcessHelper.prepopulatedDm(userId, follows);
-
-    let currentTime = momentTz().tz('America/Los_Angeles');
-    let getCurrentTimeHour = currentTime.hours();
-
-    let additionalDays = 0;
-    if (parseInt(getCurrentTimeHour) > 6) {
-      additionalDays = 1;
-    }
-    const randomTime = sample([6, 7, 8, 9]);
-    let requiredTime = momentTz()
-      .tz('America/Los_Angeles')
-      .set({hour: randomTime})
-      .add(additionalDays, 'days');
-
-    console.log('currentTime', currentTime);
-    console.log('requiredTime', requiredTime);
-
-    const diffTime = requiredTime.diff(currentTime, 'miliseconds');
-    returnPrepopulatedDm.resultPrepopulated = returnPrepopulatedDm.resultPrepopulated.filter(
-      (results) => results.targetUser.username === process.env.USERNAME_ADMIN
-    );
-
-    await automateWelcomeMsgQueue.add(returnPrepopulatedDm, {delay: diffTime});
-
+    await preAutomateWelcomeMsgProcess(returnPrepopulatedDm);
     await ProcessHelper.followUser(userId, follows);
     await ProcessHelper.followAnonymousUser(anonUserId, follows);
     await ProcessHelper.followTopic(userId, topics);
